fix: resolve sonar cube duplication issues
