+++ packages/web-service/src/pages/ecologist-experience/enter-experience/__tests__/enter-experience.spec.js
@@ -46,47 +46,6 @@ describe('The enter experience page', () => {
     })
   })
 
-  describe('getData function', () => {
-    it('returns the experience details from the api, if there have been no error on the input', async () => {
-      const request = {
-        cache: () => ({
-          getData: () => ({
-            applicationId: '26a3e94f-2280-4ea5-ad72-920d53c110fc'
-          })
-        })
-      }
-      jest.doMock('../../../../services/api-requests.js', () => ({
-        APIRequests: {
-          ECOLOGIST_EXPERIENCE: {
-            getExperienceById: jest.fn(() => ({ experienceDetails: 'experience' }))
-          }
-        }
-      }))
-      const { getData } = await import('../enter-experience.js')
-      expect(await getData(request)).toBe('experience')
-    })
-
-    it('returns the invalid input, if the user entered too much/too little', async () => {
-      let journeyData = {
-        applicationId: '26a3e94f-2280-4ea5-ad72-920d53c110fc',
-        tempInput: 'This is just test text to ensure its retrieved'
-      }
-      const request = {
-        cache: () => ({
-          getData: () => journeyData,
-          setData: (newObj) => {
-            journeyData = newObj
-          }
-        })
-      }
-      const { getData } = await import('../enter-experience.js')
-      const result = await getData(request)
-      expect(result).toEqual('This is just test text to ensure its retrieved')
-      // The tempInput prop should now be wiped
-      expect(journeyData.tempInput).toEqual(undefined)
-    })
-  })
-
   describe('the set data function', () => {
     it('stores the experience details', async () => {
       const mockPut = jest.fn()
