+++ packages/web-service/src/pages/ecologist-experience/enter-methods/__tests__/enter-methods.spec.js
@@ -45,48 +45,6 @@ describe('The enter methods page', () => {
     })
   })
 
-  describe('getData function', () => {
-    it('returns the method details from the api, if there have been no error on the input', async () => {
-      const request = {
-        cache: () => ({
-          getData: () => ({
-            applicationId: '26a3e94f-2280-4ea5-ad72-920d53c110fc'
-          })
-        })
-      }
-      jest.doMock('../../../../services/api-requests.js', () => ({
-        APIRequests: {
-          ECOLOGIST_EXPERIENCE: {
-            getExperienceById: jest.fn(() => ({ methodExperience: 'methods' }))
-          }
-        }
-      }))
-      const { getData } = await import('../enter-methods.js')
-      const result = await getData(request)
-      expect(result).toEqual('methods')
-    })
-
-    it('returns the invalid input, if the user entered too much/too little', async () => {
-      let journeyData = {
-        applicationId: '26a3e94f-2280-4ea5-ad72-920d53c110fc',
-        tempInput: 'This is just test text to ensure its retrieved'
-      }
-      const request = {
-        cache: () => ({
-          getData: () => journeyData,
-          setData: (newObj) => {
-            journeyData = newObj
-          }
-        })
-      }
-      const { getData } = await import('../enter-methods.js')
-      const result = await getData(request)
-      expect(result).toEqual('This is just test text to ensure its retrieved')
-      // The tempInput prop should now be wiped
-      expect(journeyData.tempInput).toEqual(undefined)
-    })
-  })
-
   describe('the set data function', () => {
     it('writes the method experience to the api', async () => {
       const mockPut = jest.fn()
